+++ src/Plainion.Core.Tests/Xaml/IncludeTests.cs
@@ -7,7 +7,7 @@ namespace Plainion.Tests.Xaml
     [TestFixture]
     class IncludeTests
     {
-        [Test]
+        [Test,Ignore("we currently have no idea how to specify relative paths - relative to incuding file")]
         public void Include_FileExists_XamlIncluded()
         {
             var master = Path.Combine( Path.GetDirectoryName( GetType().Assembly.Location ), "TestData/Xaml/Master.xaml" );